+++ src/Idris/Core/Elaborate.hs
@@ -463,18 +463,17 @@ equiv :: Raw -> Elab' aux ()
 equiv tm = processTactic' (Equiv tm)
 
 -- | Turn the current hole into a pattern variable with the provided
--- name, made unique if MN
+-- name, made unique if not the same as the head of the hole queue
 patvar :: Name -> Elab' aux ()
 patvar n@(SN _) = do apply (Var n) []; solve
 patvar n = do env <- get_env
               hs <- get_holes
               if (n `elem` map fst env) then do apply (Var n) []; solve
-                else do n' <- case n of
-                                    UN _ -> return $! n
-                                    MN _ _ -> unique_hole n
-                                    NS _ _ -> return $! n
-                                    x -> return $! n
-                        processTactic' (PatVar n')
+                else do n' <- case hs of
+                                   (h : hs) -> if n == h then return n
+                                                  else unique_hole n
+                                   _ -> unique_hole n
+                        processTactic' (PatVar n)
 
 -- | Turn the current hole into a pattern variable with the provided
 -- name, but don't make MNs unique.